[OU-FIX] website_sale: delete script This script should go in website, not website_sale. However, this script is causing errors in migrations to later versions and its call will be commented in case someone considers that its execution is relevant and can decomment it. The reason is that in v14 the main view website.layout was duplicated with the associated web when modifying the name of the company, something that this script reproduces and in the behaviour of later versions this view should not be duplicated. It is better not to automate tasks on main views to avoid errors in migrations to later versions.
TT44794
