+++ openupgrade_scripts/scripts/website/14.0.1.0/post-migration.py
@@ -1,8 +1,44 @@
 # Copyright 2021 ForgeFlow S.L.  <https://www.forgeflow.com>
 # Copyright 2023 Tecnativa - Pilar Vargas
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
+import re
+
 from openupgradelib import openupgrade
 from openupgradelib.openupgrade_140 import convert_html_string_13to14
+
+
+def extract_footer_copyright_company_name(env):
+    """Replace Copyright content in the footer so as not to lose
+    content from previous versions if the copyright has been customised."""
+    for website in env["website"].search([]):
+        # Search for old copyright and if it does not exist set the default company name
+        web_frontend_layout_view = env["ir.ui.view"].search(
+            [("key", "=", "web.frontend_layout"), ("website_id", "=", website.id)]
+        )
+        if web_frontend_layout_view:
+            frontend_layout_pattern = r"<span(?:.*?)>(.*?)</span>"
+            frontend_layout_matches = re.findall(
+                frontend_layout_pattern,
+                web_frontend_layout_view.arch_db,
+                re.DOTALL,
+            )
+            copyright_content = " ".join(
+                match.strip() for match in frontend_layout_matches
+            )
+        else:
+            copyright_content = f"Copyright © {website.company_id.name}"
+        # Set new copyright
+        website_layout_view = env.ref("website.layout")
+        website_layout_pattern = (
+            r'<span class="o_footer_copyright_name mr-2">(.*?)<\/span>'
+        )
+        website_layout_matches, *_ = re.findall(
+            website_layout_pattern, website_layout_view.arch_db, re.DOTALL
+        )
+        new_arch = website_layout_view.arch_db.replace(
+            website_layout_matches, copyright_content
+        )
+        website_layout_view.with_context(website_id=website.id).arch_db = new_arch
 
 
 def website_cookie_notice_post_migration(env):
@@ -44,3 +80,10 @@ def migrate(env, version):
     openupgrade.delete_records_safely_by_xml_id(env, ["website.aboutus_page"])
     website_cookie_notice_post_migration(env)
     convert_html_string(env)
+    # This call is commented out as this script causes conflicts with inheritance
+    # in migrations to later versions. In v14 a duplicate of website.layout with
+    # associated website is made when the company name is modified in the footer,
+    # this script replicates that behaviour in the migration but in later versions
+    # it can be a problem of inheritance. However, it can be uncommented and used if
+    # deemed appropriate.
+    # extract_footer_copyright_company_name(env)
